requirements.txt
absl-py==1.4.0
alabaster==0.7.13
autopep8==2.0.2
Babel==2.12.1
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
dgl==1.1.0 
-f https://data.dgl.ai/wheels/repo.html
dglgo==0.0.2
-f https://data.dgl.ai/wheels/repo.html
docutils==0.20.1
filelock==3.12.0
fonttools==4.39.4
google-auth==2.18.1
google-auth-oauthlib==1.0.0
grpcio==1.54.2
idna==3.4
imagesize==1.4.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
isort==5.12.0
jaro-winkler==2.0.3
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lit==16.0.5
littleutils==0.2.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
numpydoc==1.5.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
ogb==1.3.6
outdated==0.2.2
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==4.23.1
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pydantic==1.10.7
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
rdkit-pypi==2022.9.5
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
snowballstemmer==2.2.0
sphinx==7.0.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==2.0.1
torch-tb-profiler==0.4.1
torchmetrics==0.11.4
tqdm==4.65.0
triton==2.0.0
typer==0.9.0
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.4
zipp==3.15.0
